requirements.txt
attrs==25.3.0
black==25.1.0
boto3==1.40.11
botocore==1.40.11
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.3
cryptography==45.0.6
durationpy==0.10
flake8==7.3.0
google-auth==2.40.3
h11==0.16.0
hvac==2.3.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
kubernetes==33.1.0
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.2
oauthlib==3.3.1
openshift==0.13.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-string-utils==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
structlog==25.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yamllint==1.34.0